Fix the use of jqueryui-touch-punch.
Jqueryui-touch-punch is necessary for mobile devices.

There are a lot of bower packages to jquery-touch-punch
(try to look the github network for this project).
I've chosen the most stared on bower.io.
However, there is no version on this repository so, we cannot fix the
version and we just hope that nothing will break in the future...